convert integers to smaller integers
Former-commit-id: 9f037f4ef10692c04d506a7db163b045374095f2

+++ src/projection/utils/coo_matrix_nocheck.py
@@ -2,7 +2,7 @@ import numpy as np
 
 class CooMatrix:
     def __init__(self, row, col, data, shape):
-        self.row = np.array(row)
-        self.col = np.array(col)
+        self.row = np.array(row, dtype=np.uint32)
+        self.col = np.array(col, dtype=np.uint32)
         self.data = np.array(data)
         self.shape = shape
